@@ -1,13 +1,16 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import { Redirect } from 'react-router-dom'
 import decode from 'jwt-decode';
+import { getNotes } from '../../functions/notes';
 
-class Home extends React.Component {
-  login = () => {
+function Home() {
+  const [notes, setNotes] = useState([]);
+
+  const login = () => {
     this.props.history.push('/login')
   }
 
-  checkAdmin = () => {
+  const checkAdmin = () => {
     const token = sessionStorage.getItem('token');
     if(!token) return false;
     const decoded = decode(token);
@@ -16,16 +19,26 @@ class Home extends React.Component {
     return true;
   }
 
-  render() {
-    return this.checkAdmin() ? (
-      <Redirect to={{ pathname: '/admin' }} />
-    ) : (
-      <div>
-        <h1>Home</h1>
-        <p><button onClick={this.login}>Logout</button></p>
-      </div>
-    )
-  }
+  useEffect(() => {
+    let mounted = true;
+    getNotes()
+      .then(items => {
+        console.log(items);
+        if (mounted) {
+          setNotes(items)
+        }
+      })
+    return () => mounted = false;
+  }, [])
+
+  return checkAdmin() ? (
+    <Redirect to={{ pathname: '/admin' }} />
+  ) : (
+    <div>
+      <h1>Home</h1>
+      <p><button onClick={login}>Logout</button></p>
+    </div>
+  )
 }
 
 export default Home
